Seperate multiple http2 handler and request handler
